% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{datasets.download_files}
\alias{datasets.download_files}
\title{Download a Foundry File locally.}
\usage{
datasets.download_files(alias, file_paths)
}
\arguments{
\item{alias}{The alias representing the Dataset.}

\item{file_paths}{The file paths to download from the alias.}
}
\value{
The local file paths where files were downloaded.
}
\description{
Download a Foundry File locally.
}
\examples{
\dontrun{
# download all files in dataset
all_files <- datasets.list_files("my_alias")
downloaded_files <- datasets.download_files("my_alias", all_files)
}
}
