% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{datasets.list_files}
\alias{datasets.list_files}
\title{Lists the files stored in a Foundry Dataset.}
\usage{
datasets.list_files(alias)
}
\arguments{
\item{alias}{The alias representing the Dataset.}
}
\value{
The lists of file properties.
}
\description{
Lists the files stored in a Foundry Dataset.
}
\examples{
\dontrun{
# Get all PDF files in dataset
all_files <- datasets.list_files("my_dataset")
pdf_files <- all_files[sapply(all_files, function(x) grepl(".*\\\\.pdf", x$path))]

# Get all file names
file_names <- sapply(all_files, function(x) x$path)
}
}
