% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{datasets.write_table}
\alias{datasets.write_table}
\title{Writes a data.frame to a Foundry dataset.}
\usage{
datasets.write_table(data, alias)
}
\arguments{
\item{data}{A data.frame or an arrow Table.}

\item{alias}{The alias representing the Dataset.}
}
\description{
Note that types may not be exactly preserved and all types are not supported.
See https://arrow.apache.org/docs/r/articles/arrow.html for details on type conversions
from a data.frame to an arrow Table. Use arrow::arrow_table to use more granular types.
}
