% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{download_files}
\alias{download_files}
\title{Download multiple files from a Foundry Dataset locally.}
\usage{
download_files(dataset, target, path = NULL, ...)
}
\arguments{
\item{dataset}{A Dataset or a character string representing the RID or dataset path.}

\item{target}{A character string representing the directory where the file should be downloaded.
The directory or its parent should exist.}

\item{path}{If present, only the specified file or the files in the specified folder will be downloaded.}

\item{...}{Extra parameters to pass to the download.file() function.}
}
\description{
Download multiple files from a Foundry Dataset locally.
}
