% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_config}
\alias{get_config}
\title{Loads a config from an environment variable with format `FOUNDRY_CONFIG_KEY`
or from an option with format `foundry.config.key`.}
\usage{
get_config(name, default = NULL)
}
\description{
Loads a config from an environment variable with format `FOUNDRY_CONFIG_KEY`
or from an option with format `foundry.config.key`.
}
\keyword{internal}
