% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_dataset}
\alias{get_dataset}
\title{Retrieves a Foundry Dataset.}
\usage{
get_dataset(
  dataset_ref,
  branch = NULL,
  transaction_range = NULL,
  create = FALSE
)
}
\arguments{
\item{dataset_ref}{A dataset RID or path.}

\item{branch}{The dataset branch.}

\item{transaction_range}{A list of 2 nullable characters representing the start and end transaction RIDs.}

\item{create}{A boolean indicating whether the dataset should be created if it does not exist yet.}
}
\value{
A Dataset object representing the dataset view.
}
\description{
Returns an object representing a specific view of a Foundry dataset.
}
