% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_palantir}
\alias{install_palantir}
\title{Installs pypalantir}
\usage{
install_palantir(
  version = "0.8.0",
  python_version = "3.9",
  envname = "r-reticulate",
  restart_session = TRUE,
  recreate_environment = FALSE
)
}
\arguments{
\item{version}{Palantir Python SDK version (https://github.com/palantir/palantir-python-sdk)}

\item{python_version}{Version of python}

\item{envname}{Conda environment name}

\item{restart_session}{Whether the R session should be restarted to ensure the python package is updated}

\item{recreate_environment}{Whether to recreate the conda environment from scratch}
}
\description{
Installs the palantir python package and required dependencies. If Miniconda is not
installed yet, you will be prompted to install it first. We recommend restarting the
R session after calling this function.
}
