% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{list_files}
\alias{list_files}
\title{Lists the files stored in a Foundry Dataset.}
\usage{
list_files(
  dataset_rid,
  branch_id = NULL,
  start_transaction_rid = NULL,
  end_transaction_rid = NULL,
  limit = NULL
)
}
\arguments{
\item{dataset_rid}{The RID of the Dataset.}

\item{branch_id}{The identifier (name) of the Branch.}

\item{start_transaction_rid}{The Resource Identifier (RID) of the start Transaction.}

\item{end_transaction_rid}{The Resource Identifier (RID) of the end Transaction.}

\item{limit}{The maximum number of files to return.}
}
\value{
A list of File objects representing the files in the dataset.
}
\description{
Lists the files stored in a Foundry Dataset.
}
