% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{read_table}
\alias{read_table}
\title{Reads a tabular Foundry dataset as data.frame.}
\usage{
read_table(
  dataset_rid,
  branch_id = NULL,
  start_transaction_rid = NULL,
  end_transaction_rid = NULL,
  columns = NULL,
  row_limit = NULL
)
}
\arguments{
\item{dataset_rid}{The RID of the Dataset. It must be tabular, i.e. have a schema.}

\item{branch_id}{The identifier (name) of the Branch.}

\item{start_transaction_rid}{The Resource Identifier (RID) of the start Transaction.}

\item{end_transaction_rid}{The Resource Identifier (RID) of the end Transaction.}
}
\value{
A data.table
}
\description{
Note that types may not match exactly the Foundry column types.
See https://arrow.apache.org/docs/r/articles/arrow.html for details on type conversions
from an arrow Table to a data.frame.
For more advanced usage, use `read_table_arrow`.
}
