% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{read_table_arrow}
\alias{read_table_arrow}
\title{Reads a tabular Foundry dataset as an arrow Table.}
\usage{
read_table_arrow(dataset)
}
\arguments{
\item{dataset}{A Dataset or a character string representing the RID or dataset path.
The Foundry dataset must be tabular, i.e. have a schema.}
}
\value{
An arrow Table
}
\description{
The `arrow` library must be loaded prior to calling this function.
}
\examples{
\dontrun{
library(arrow)
library(dplyr)

df <- read_table.arrow("/path/to/dataset")
    \%>\% select("column" == "value")
    \%>\% collect()
}
}
