% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{write_table}
\alias{write_table}
\title{Writes a data.frame to a Foundry dataset.}
\usage{
write_table(data, dataset, branch = NULL)
}
\arguments{
\item{data}{A data.frame or an arrow Table.}

\item{dataset}{A Dataset or a character string representing the RID or dataset path.}

\item{branch}{The dataset branch.}
}
\description{
Note that types may not be exactly preserved and all types are not supported.
See https://arrow.apache.org/docs/r/articles/arrow.html for details on type conversions
from a data.frame to an arrow Table. Use arrow::arrow_table to use more granular types.
}
