% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{datasets.list_files}
\alias{datasets.list_files}
\title{Lists the files stored in a Foundry Dataset.}
\usage{
datasets.list_files(alias, pattern = NULL, path_only = TRUE)
}
\arguments{
\item{alias}{The alias representing the Dataset.}

\item{pattern}{A regex to filter file paths on.}

\item{path_only}{If true, return the file paths only, otherwise return lists with all file properties.}
}
\value{
The file paths or lists of file properties.
}
\description{
Lists the files stored in a Foundry Dataset.
}
