% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{datasets.list_files}
\alias{datasets.list_files}
\title{Lists the files stored in a Foundry Dataset.}
\usage{
datasets.list_files(alias, regex = ".*")
}
\arguments{
\item{alias}{The alias representing the Dataset.}

\item{regex}{A regex used to filter files by path.}
}
\value{
The lists of file properties.
}
\description{
Lists the files stored in a Foundry Dataset.
}
\examples{
\dontrun{
# List all PDF files in a Dataset
all_files <- datasets.list_files("my_dataset", regex=".*\\\\.pdf")

# Get all file names
file_names <- sapply(all_files, function(x) x$path)
}

}
