% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{datasets.upload_files}
\alias{datasets.upload_files}
\title{Upload a local file or folder to a Foundry Dataset.}
\usage{
datasets.upload_files(local_files, alias)
}
\arguments{
\item{local_files}{The files and folders to upload.
If a folder is provided, all files found recursively in subfolders will be uploaded.}

\item{alias}{The alias representing the Dataset.}
}
\value{
The list of local file paths and corresponding file name in the Foundry dataset.
}
\description{
Upload a local file or folder to a Foundry Dataset.
}
