% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{download_file}
\alias{download_file}
\title{Download a Foundry File locally.}
\usage{
download_file(
  dataset_rid,
  file_path,
  target,
  branch_id = NULL,
  start_transaction_rid = NULL,
  end_transaction_rid = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{dataset_rid}{The RID of the Dataset.}

\item{target}{A character string representing the location where the file should be downloaded.}

\item{branch_id}{The identifier (name) of the Branch.}

\item{start_transaction_rid}{The Resource Identifier (RID) of the start Transaction.}

\item{end_transaction_rid}{The Resource Identifier (RID) of the end Transaction.}

\item{overwrite}{whether the target should be overwritten.}

\item{logical_path}{A character string representing the logical path of the file in the Dataset.}
}
\description{
Download a Foundry File locally.
}
