% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{download_file}
\alias{download_file}
\title{Download a Foundry File locally.}
\usage{
download_file(
  dataset,
  logical_path,
  target,
  ...,
  branch = NULL,
  transaction = NULL
)
}
\arguments{
\item{dataset}{A Dataset or a character string representing the RID or dataset path.}

\item{logical_path}{A character string representing the logical path of the file in the Dataset.}

\item{target}{A character string representing the location where the file should be downloaded.}

\item{...}{Extra parameters to pass to the download.file() function.}

\item{branch}{The dataset branch.}

\item{transaction}{The dataset transaction.}
}
\description{
Download a Foundry File locally.
}
