% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{download_files}
\alias{download_files}
\title{Download multiple files from a Foundry Dataset locally.}
\usage{
download_files(
  dataset,
  target,
  ...,
  branch = NULL,
  transaction = NULL,
  path = NULL
)
}
\arguments{
\item{dataset}{A Dataset or a character string representing the RID or dataset path.}

\item{target}{A character string representing the directory where the file should be downloaded.
The directory or its parent should exist.}

\item{...}{Extra parameters to pass to the download.file() function.}

\item{branch}{The dataset branch.}

\item{transaction}{The dataset transaction.}

\item{path}{If present, only the file or the files with the specified prefix will be downloaded.}
}
\description{
Download multiple files from a Foundry Dataset locally.
}
