% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{download_files}
\alias{download_files}
\title{Download multiple files from a Foundry Dataset locally.}
\usage{
download_files(
  dataset_rid,
  target,
  branch_id = NULL,
  start_transaction_rid = NULL,
  end_transaction_rid = NULL
)
}
\arguments{
\item{dataset_rid}{The RID of the Dataset.}

\item{target}{A character string representing the directory where the file should be downloaded.
The directory or its parent should exist. If it exists, the directory must be empty.}

\item{branch_id}{The identifier (name) of the Branch.}

\item{start_transaction_rid}{The Resource Identifier (RID) of the start Transaction.}

\item{end_transaction_rid}{The Resource Identifier (RID) of the end Transaction.}
}
\description{
Download multiple files from a Foundry Dataset locally.
}
