% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{list_files}
\alias{list_files}
\title{Lists the files stored in a Foundry Dataset.}
\usage{
list_files(dataset, path = NULL)
}
\arguments{
\item{dataset}{A Dataset or a character string representing the RID or dataset path.}

\item{path}{If present, only the specified file or the files in the specified folder will be returned.}
}
\value{
A list of File objects representing the files in the dataset.
}
\description{
Lists the files stored in a Foundry Dataset.
}
