% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{read_table}
\alias{read_table}
\title{Reads a tabular Foundry dataset as data.frame.}
\usage{
read_table(dataset)
}
\arguments{
\item{dataset}{A Dataset or a character string representing the RID or dataset path.
The Foundry dataset must be tabular, i.e. have a schema.}
}
\value{
A data.table
}
\description{
Note that the variable classes may not match exactly the Foundry column types.
A STRING column with NULL values will be cast as a list variable to properly
capture NULL values and empty strings.
For more advanced usage, use `read_table_arrow`.
}
