% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{read_table}
\alias{read_table}
\title{Reads a tabular Foundry dataset as data.frame.}
\usage{
read_table(dataset, branch = NULL, transaction = NULL)
}
\arguments{
\item{dataset}{A Dataset or a character string representing the RID or dataset path.
The Foundry dataset must be tabular, i.e. have a schema.}

\item{branch}{The dataset branch.}

\item{transaction}{The dataset transaction.}
}
\value{
A data.table
}
\description{
Note that types may not match exactly the Foundry column types.
See https://arrow.apache.org/docs/r/articles/arrow.html for details on type conversions
from an arrow Table to a data.frame.
For more advanced usage, use `read_table_arrow`.
}
