% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{read_table_arrow}
\alias{read_table_arrow}
\title{Reads a tabular Foundry dataset as an arrow Table.}
\usage{
read_table_arrow(
  dataset_rid,
  branch_id = NULL,
  start_transaction_rid = NULL,
  end_transaction_rid = NULL,
  columns = NULL,
  row_limit = NULL
)
}
\arguments{
\item{dataset_rid}{The RID of the Dataset. It must be tabular, i.e. have a schema.}

\item{branch_id}{The identifier (name) of the Branch.}

\item{start_transaction_rid}{The Resource Identifier (RID) of the start Transaction.}

\item{end_transaction_rid}{The Resource Identifier (RID) of the end Transaction.}
}
\value{
An arrow Table
}
\description{
Reads a tabular Foundry dataset as an arrow Table.
}
