% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{read_table_arrow}
\alias{read_table_arrow}
\title{Reads a tabular Foundry dataset as an arrow Table.}
\usage{
read_table_arrow(
  dataset,
  ...,
  branch = NULL,
  transaction = NULL,
  timeout = 150
)
}
\arguments{
\item{dataset}{A Dataset or a character string representing the RID or dataset path.
The Foundry dataset must be tabular, i.e. have a schema.}

\item{...}{Ignored, used to bind optional arguments by name}

\item{branch}{The dataset branch.}

\item{transaction}{The dataset transaction.}

\item{timeout}{The request timeout, in seconds.}
}
\value{
An arrow Table
}
\description{
The `arrow` library must be loaded prior to calling this function.
}
\examples{
\dontrun{
library(arrow)
library(dplyr)

df <- read_table_arrow("/path/to/dataset")
    \%>\% select("column" == "value")
    \%>\% collect()
}
}
