% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{write_table}
\alias{write_table}
\title{Writes a data.frame to a Foundry dataset.}
\usage{
write_table(data, dataset)
}
\arguments{
\item{data}{A data.frame.}

\item{dataset}{A Dataset or a character string representing the RID or dataset path.}
}
\description{
Writes a data.frame to a Foundry dataset.
}
