% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{write_table}
\alias{write_table}
\title{Writes a data.frame to a Foundry dataset.}
\usage{
write_table(data, dataset_rid, branch_id = NULL)
}
\arguments{
\item{data}{A data.frame or an arrow Table.}

\item{dataset_rid}{The RID of the Dataset.}

\item{branch_id}{The identifier (name) of the Branch.}
}
\description{
Note that types may not be exactly preserved and all types are not supported.
See https://arrow.apache.org/docs/r/articles/arrow.html for details on type conversions
from a data.frame to an arrow Table. Use arrow::arrow_table to use more granular types.
}
